fix yandex OCR api parser

+++ src/main/java/ru/ssugt/capture/SetRectangle.java
@@ -3,6 +3,6 @@ package ru.ssugt.capture;
 import lombok.Getter;
 
 
-public record SetRectangle(double x, double y, double width, double height) {
+public record SetRectangle(int x, int y, int width, int height) {
 
 }
